requirements.txt
pandas>=1.1.4
numpy>=1.18.5
scikit-learn
pymongo==4.3.3
fastapi==0.110.2
uvicorn==0.20.0
python-multipart
dnspython==2.2.1
certifi==2019.9.11
gunicorn
langchain==0.1.0
openai==1.7.2
python-dotenv==1.0.0
pinecone-client==2.2.4
langchain-openai==0.0.2.post1
chromadb
faiss-cpu==1.7.4
pycryptodome==3.15.0
openapi-schema-pydantic==1.2.4
tiktoken==0.5.2
python-multipart
pyzbar
opencv-python-headless
replicate
pillow
yagmail[all]
azure-storage-blob
streamlit
uuid